+++ packages/manager/src/features/LoadBalancers/LoadBalancerDetail/Certificates/Certificates.tsx
@@ -3,6 +3,7 @@ import React, { useState } from 'react';
 import { useHistory, useLocation, useParams } from 'react-router-dom';
 
 import { ActionMenu } from 'src/components/ActionMenu';
+import { Button } from 'src/components/Button/Button';
 import { CircleProgress } from 'src/components/CircleProgress';
 import { Link } from 'src/components/Link';
 import { PaginationFooter } from 'src/components/PaginationFooter/PaginationFooter';
@@ -31,7 +32,6 @@ export const Certificates = () => {
     loadbalancerId: string;
     type: Certificate['type'] | undefined;
   }>();
-
   const certType = type ?? 'downstream';
 
   const id = Number(loadbalancerId);
@@ -84,27 +84,46 @@ export const Certificates = () => {
 
   return (
     <Stack paddingTop={1} spacing={2}>
-      {certType === 'ca' ? (
-        <Typography>
-          Used by the load balancer to accept responses from your endpoints in
-          your Service Target. This is the certificate installed on your
-          Endpoints. Apply these certificate(s) in the{' '}
-          <Link to={`/loadbalancers/${loadbalancerId}/service-targets`}>
-            Service Targets
-          </Link>{' '}
-          Tab.
-        </Typography>
-      ) : (
-        <Typography>
-          Certificate used by your load balancer to terminate the connection and
-          decrypt request from client. Apply these certificate(s) in the Details
-          section of your HTTPS{' '}
-          <Link to={`/loadbalancers/${loadbalancerId}/configurations`}>
-            Configuration
-          </Link>
-          .
-        </Typography>
-      )}
+      <Stack
+        alignItems="center"
+        direction="row"
+        flexWrap={{ md: 'unset', sm: 'wrap', xs: 'wrap' }}
+        gap={2}
+        justifyContent="space-between"
+      >
+        {certType === 'ca' ? (
+          <Typography>
+            Used by the load balancer to accept responses from your endpoints in
+            your Service Target. This is the certificate installed on your
+            Endpoints. Apply these certificate(s) in the{' '}
+            <Link to={`/loadbalancers/${loadbalancerId}/service-targets`}>
+              Service Targets
+            </Link>{' '}
+            Tab.
+          </Typography>
+        ) : (
+          <Typography>
+            Certificate used by your load balancer to terminate the connection
+            and decrypt request from client. Apply these certificate(s) in the
+            Details section of your HTTPS{' '}
+            <Link to={`/loadbalancers/${loadbalancerId}/configurations`}>
+              Configuration
+            </Link>
+            .
+          </Typography>
+        )}
+        <Button
+          onClick={() =>
+            history.push(
+              `/loadbalancers/${loadbalancerId}/certificates/${certType}/create`
+            )
+          }
+          buttonType="primary"
+          sx={{ minWidth: 200 }}
+        >
+          Upload Certificate
+        </Button>
+      </Stack>
       <Table>
         <TableHead>
           <TableRow>
