+++ packages/manager/src/features/LoadBalancers/LoadBalancerDetail/Certificates/index.tsx
@@ -1,14 +1,7 @@
 import Stack from '@mui/material/Stack';
 import React from 'react';
-import {
-  Route,
-  useHistory,
-  useLocation,
-  useRouteMatch,
-} from 'react-router-dom';
+import { Route, useLocation, useRouteMatch } from 'react-router-dom';
 
-import { Box } from 'src/components/Box';
-import { Button } from 'src/components/Button/Button';
 import { Tabs } from 'src/components/ReachTabs';
 import { SuspenseLoader } from 'src/components/SuspenseLoader';
 import { TabLinkList } from 'src/components/TabLinkList/TabLinkList';
@@ -23,7 +16,6 @@ const Certificates = React.lazy(() =>
 export const LoadBalancerCertificates = () => {
   const { path, url } = useRouteMatch();
   const location = useLocation();
-  const history = useHistory();
 
   const tabs = [
     {
@@ -41,20 +33,11 @@ export const LoadBalancerCertificates = () => {
   );
 
   return (
-    <>
-      <Stack alignItems="center" direction="row" justifyContent="space-between">
-        <Typography>
-          Upload certificates to your Load Balancer for use across your
-          Configurations and Service Targets.
-        </Typography>
-        <Box flexGrow={1} />
-        <Button
-          buttonType="primary"
-          onClick={() => history.push(location.pathname + '/create')}
-        >
-          Upload Certificate
-        </Button>
-      </Stack>
+    <Stack paddingTop={1} spacing={1}>
+      <Typography>
+        Upload certificates to your Load Balancer for use across your
+        Configurations and Service Targets.
+      </Typography>
       <Tabs index={tabIndex === -1 ? 0 : tabIndex}>
         <TabLinkList
           tabs={tabs.map((t) => ({ ...t, routeName: `${url}/${t.path}` }))}
@@ -63,6 +46,6 @@ export const LoadBalancerCertificates = () => {
           <Route component={Certificates} path={`${path}/:type?/:create?`} />
         </React.Suspense>
       </Tabs>
-    </>
+    </Stack>
   );
 };
